Move content serialization + subset from importer to worker
